requirements.txt
flask==3.1.2
wordcloud==1.9.4
pillow==12.0.0
numpy==2.3.5